pyproject.toml
[project]
name = "jarvis-mini-server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.1",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "spotipy>=2.26.0",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "ruff>=0.15.12",
]
